+++ src/App.js
@@ -1,18 +1,18 @@
 import React, { useState, useEffect } from "react";
-import Preloader from "../src/components/Pre";
-import Navbar from "./components/Navbar";
-import Home from "./components/Home/Home";
-import About from "./components/About/About";
-import Projects from "./components/Projects/Projects";
-import Footer from "./components/Footer";
-import Resume from "./components/Resume/ResumeNew";
+import Preloader from "./components/core/Pre";
+import Navbar from "./components/core/Navbar";
+import Home from "./containers/Home/Home";
+import About from "./containers/About/About";
+import Projects from "./containers/Projects/Projects";
+import Footer from "./components/core/Footer";
+import Resume from "./containers/Resume/ResumeNew";
 import {
   BrowserRouter as Router,
   Route,
   Routes,
   Navigate
 } from "react-router-dom";
-import ScrollToTop from "./components/ScrollToTop";
+import ScrollToTop from "./components/core/ScrollToTop";
 import "./style.css";
 import "./App.css";
 import "bootstrap/dist/css/bootstrap.min.css";
