+++ src/containers/Home/Home.js
@@ -1,9 +1,8 @@
 import React from "react";
 import { Container, Row, Col } from "react-bootstrap";
-import homeLogo from "../../Assets/home-main.svg";
-import Particle from "../Particle";
-import Home2 from "./Home2";
-import Type from "./Type";
+import homeLogo from "../../assets/home-main.svg";
+import Particle from "../../components/core/Particle";
+import Type from "../../components/home/Type";
 
 function Home() {
   return (
@@ -22,7 +21,7 @@ function Home() {
 
               <h1 className="heading-name">
                 I'M
-                <strong className="main-name"> SOUMYAJIT BEHERA</strong>
+                <strong className="main-name"> María Hernández</strong>
               </h1>
 
               <div style={{ padding: 50, textAlign: "left" }}>
@@ -41,7 +40,6 @@ function Home() {
           </Row>
         </Container>
       </Container>
-      <Home2 />
     </section>
   );
 }
