+++ src/containers/Projects/Projects.js
@@ -1,13 +1,13 @@
 import React from "react";
 import { Container, Row, Col } from "react-bootstrap";
-import ProjectCard from "./ProjectCards";
-import Particle from "../Particle";
-import leaf from "../../Assets/Projects/leaf.png";
-import emotion from "../../Assets/Projects/emotion.png";
-import editor from "../../Assets/Projects/codeEditor.png";
-import chatify from "../../Assets/Projects/chatify.png";
-import suicide from "../../Assets/Projects/suicide.png";
-import bitsOfCode from "../../Assets/Projects/blog.png";
+import ProjectCard from "../../components/project/ProjectCards";
+import Particle from "../../components/core/Particle";
+import leaf from "../../assets/Projects/leaf.png";
+import emotion from "../../assets/Projects/emotion.png";
+import editor from "../../assets/Projects/codeEditor.png";
+import chatify from "../../assets/Projects/chatify.png";
+import suicide from "../../assets/Projects/suicide.png";
+import bitsOfCode from "../../assets/Projects/blog.png";
 
 function Projects() {
   return (
